+++ src/app/(pages)/(search)/layout.tsx
@@ -1,0 +1,15 @@
+import { SearchBar } from '@/app/components/features/search/SearchBar';
+
+export const metadata = {
+  title: 'Home',
+  description: 'Generated by create next app',
+};
+
+export default function SearchLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <div className='flex flex-col'>
+      <SearchBar />
+      {children}
+    </div>
+  );
+}
