+++ src/app/(pages)/(search)/page.tsx
@@ -1,0 +1,12 @@
+import { RecipeListHorizontal, RecipeListVertical } from '@/app/components/features/recipe/RecipeList';
+
+export default function Home() {
+  return (
+    <main>
+      <div className='h-8'/> {/* 後で削除 */}
+      <RecipeListHorizontal recipeList={[]} />
+      <div className='h-8'/> {/* 後で削除 */}
+      <RecipeListVertical recipeList={[]} />
+    </main>
+  );
+}
